+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #4 - entity BankCard was created, integrated into gazon-service and tested
 - #3 - created passport entity
 - #7 - entities Product and ProductImage were created and integrated
+- #10 - entity SelectedProduct was created, but not fully integrated
 
 ## [1.0.0]